Ensure that MultiQC runs when skipBiotypeQC
If the user has `skipBiotypeQC = true`, the featureCounts process will not dump anything into the featureCounts_biotype channel (`optional true into featureCounts_biotype`). This empty channel will prevent the multiqc process from running, as there is no ifEmpty condition applied to the featureCounts_biotype channel in the multiqc process inputs.
This should fix this problem for users who are not necessarily interested in the biotype counts.